add code to send out the 4th/5th emails
