+++ src/main/scala/com/gu/effects/Logging.scala
@@ -1,4 +1,4 @@
-package com.gu.util
+package com.gu.effects
 
 import org.apache.log4j.Logger
