+++ src/main/scala/com/gu/util/apigateway/ApiGatewayHandler.scala
@@ -3,14 +3,15 @@ package com.gu.util.apigateway
 import java.io.{ InputStream, OutputStream }
 
 import com.amazonaws.services.lambda.runtime.Context
+import com.gu.effects.{ Logging, StateHttp }
 import com.gu.util.Auth.{ credentialsAreValid, validTenant }
 import com.gu.util._
 import com.gu.util.apigateway.ApiGatewayResponse.{ outputForAPIGateway, successfulExecution, unauthorized }
 import com.gu.util.apigateway.ResponseModels.ApiResponse
 import com.gu.util.zuora.Types._
-import com.gu.util.zuora.ZuoraRestRequestMaker
 import play.api.libs.json.Json
 
+import scala.io.Source
 import scala.util.Try
 import scalaz.{ -\/, \/, \/- }
 
@@ -20,7 +21,7 @@ object ApiGatewayHandler extends Logging {
     configAttempt: Try[Config],
     stage: String,
 
-    getZuoraRestService: Try[ZuoraRestRequestMaker],
+    getZuoraRestService: Try[StateHttp],
     operation: (ApiGatewayRequest, Config) => ZuoraOp[Unit]
   )
 
@@ -29,9 +30,11 @@ object ApiGatewayHandler extends Logging {
   */
   def handleRequest(inputStream: InputStream, outputStream: OutputStream, context: Context)(lambdaConfig: LambdaConfig): Unit = {
     logger.info(s"${this.getClass} Lambda is starting up in ${lambdaConfig.stage}")
+    val jsonString = Source.fromInputStream(inputStream).mkString
+    logger.info(s"payload from api gateway is: $jsonString")
     val response = for {
       config <- lambdaConfig.configAttempt.toFailableOp("load config")
-      apiGatewayRequest <- Json.parse(inputStream).validate[ApiGatewayRequest].toFailableOp
+      apiGatewayRequest <- Json.parse(jsonString).validate[ApiGatewayRequest].toFailableOp
       auth <- authenticateCallout(apiGatewayRequest.requestAuth, config.trustedApiConfig)
       _ = logger.info("Authenticated request successfully...")
       _ = logger.info(s"Body from Zuora was: ${apiGatewayRequest.body}")
