+++ src/main/scala/com/gu/util/apigateway/ApiGatewayResponse.scala
@@ -2,7 +2,7 @@ package com.gu.util.apigateway
 
 import java.io.{ OutputStream, OutputStreamWriter }
 
-import com.gu.util.Logging
+import com.gu.effects.Logging
 import com.gu.util.apigateway.ResponseModels.{ ApiResponse, Headers }
 import com.gu.util.apigateway.ResponseWriters._
 import play.api.libs.json.{ Json, Writes }
