+++ src/test/scala/com/gu/paymentFailure/PaymentFailureHandlerTest.scala
@@ -4,7 +4,8 @@ import java.io.ByteArrayOutputStream
 
 import com.amazonaws.services.sqs.model.SendMessageResult
 import com.gu.autoCancel.AutoCancelHandler
-import com.gu.util.exacttarget.EmailClient.SendDataExtensionToQueue
+import com.gu.effects.StateHttp
+import com.gu.util.exacttarget.EmailClient.SendEmail
 import com.gu.util.apigateway.ApiGatewayResponse.unauthorized
 import com.gu.util.apigateway.ResponseModels.ApiResponse
 import com.gu.util.zuora.Types.{ FailableOp, ZuoraOp, ZuoraReader }
@@ -14,7 +15,6 @@ import com.gu.util._
 import com.gu.util.apigateway.ApiGatewayHandler.LambdaConfig
 import com.gu.util.apigateway.{ ApiGatewayHandler, ApiGatewayResponse }
 import com.gu.util.exacttarget._
-import com.gu.util.zuora.ZuoraRestRequestMaker
 import org.joda.time.LocalDate
 import org.mockito.Matchers.any
 import org.mockito.Mockito._
@@ -43,28 +43,17 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
 
   def lambdaConfig(
     fakeInvoiceTransactionSummary: InvoiceTransactionSummary = basicInvoiceTransactionSummary,
-    etClient: SendDataExtensionToQueue = req => EitherT[ZuoraReader, ApiResponse, Unit](Reader { zhttp => ??? })
+    etClient: SendEmail = req => EitherT[ZuoraReader, ApiResponse, Unit](Reader { zhttp => ??? })
   ) = {
 
     def stage: String = "DEV"
     def configAttempt: Try[Config] = Success(Config(fakeApiConfig, fakeZuoraConfig, fakeETConfig))
-    //    def queueClient: QueueClient = fakeQueueClient
-    def getZuoraRestService: Try[ZuoraRestRequestMaker] = Success(mock[ZuoraRestRequestMaker]) //needed? no but needs further refactoring
+    def getZuoraRestService: Try[StateHttp] = Success(mock[StateHttp]) //needed? no but needs further refactoring
 
     def getInvoiceTransactions: GetInvoiceTransactions = _ => ZuoraOp.lift(fakeInvoiceTransactionSummary)
 
     LambdaConfig(configAttempt, stage, getZuoraRestService, PaymentFailureSteps.performZuoraAction(etClient, getInvoiceTransactions))
-    //    ApiGatewayHandler.handleRequest(inputStream, outputStream, context)(lambdaConfig)
   }
-
-  //  def lambda(fakeInvoiceTransactionSummary: InvoiceTransactionSummary = basicInvoiceTransactionSummary) = new PaymentFailureLambda {
-  //    override def stage: String = "DEV"
-  //    override def configAttempt: Try[Config] = Success(Config(fakeApiConfig, fakeZuoraConfig))
-  //    override def queueClient: QueueClient = fakeQueueClient
-  //    override def getZuoraRestService: Try[ZuoraRestService] = Success(mock[ZuoraRestService]) //needed? no but needs further refactoring
-  //
-  //    override def getInvoiceTransactions: GetInvoiceTransactions = _ => ZuoraOp.lift(fakeInvoiceTransactionSummary)
-  //  }
 
   val missingCredentialsResponse = """{"statusCode":"401","headers":{"Content-Type":"application/json"},"body":"Credentials are missing or invalid"}"""
   val successfulResponse = """{"statusCode":"200","headers":{"Content-Type":"application/json"},"body":"Success"}"""
@@ -115,8 +104,6 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
     var storedReq: Option[EmailRequest] = None
 
     val os = new ByteArrayOutputStream()
-    //    when(fakeZuoraRest.getInvoiceTransactions("accountId")).thenReturn(\/-(basicInvoiceTransactionSummary))
-    //    when(fakeQueueClient.sendDataExtensionToQueue(any[Message])).thenReturn(Success(mock[SendMessageResult]))
     //execute
     ApiGatewayHandler.handleRequest(stream, os, null)(lambdaConfig(
       fakeInvoiceTransactionSummary = basicInvoiceTransactionSummary,
@@ -127,7 +114,7 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
     ))
 
     //verify
-    val responseString = new String(os.toByteArray(), "UTF-8")
+    val responseString = new String(os.toByteArray, "UTF-8")
 
     val expectedMessage = EmailRequest(
       attempt = 1,
@@ -157,7 +144,6 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
       )
     )
 
-    //    verify(fakeQueueClient).sendDataExtensionToQueue(expectedMessage)
     storedReq should be(Some(expectedMessage))
     responseString jsonMatches successfulResponse
   }
@@ -167,10 +153,6 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
     val stream = getClass.getResourceAsStream("/paymentFailure/validRequest.json")
     val output = new ByteArrayOutputStream
     val invoiceTransactionSummary = InvoiceTransactionSummary(List())
-
-    //    when(fakeZuoraRest.getInvoiceTransactions(accountId)).thenReturn(\/-(invoiceTransactionSummary))
-
-    //    when(fakeQueueClient.sendDataExtensionToQueue(any[Message])).thenReturn(Success(mock[SendMessageResult]))
 
     val os = new ByteArrayOutputStream()
 
@@ -190,11 +172,6 @@ class PaymentFailureHandlerTest extends FlatSpec with MockitoSugar {
     val output = new ByteArrayOutputStream
 
     var storedReq: Option[EmailRequest] = None
-    //    when(fakeZuoraRest.getInvoiceTransactions("accountId")).thenReturn(\/-(basicInvoiceTransactionSummary))
-
-    //    when(fakeQueueClient.sendDataExtensionToQueue(any[Message])).thenReturn(Success(mock[SendMessageResult]))
-    //    when(fakeQueueClient.sendDataExtensionToQueue(any[Message])).thenReturn(Failure(new Exception("something failed!")))
-
     val os = new ByteArrayOutputStream()
 
     //execute
